+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added Calendar Filter sub-component.
+
 ## 4.43.0 - (August 4, 2020)
 
 * Changed
@@ -813,4 +816,4 @@
 
 ## 1.0.0 - (June 28, 2017)
 
-Initial stable release
+* Initial stable release